chore: rename, use own date-format utility everywhere
