+++ client/components/Admin/Share/ShareList.tsx
@@ -4,7 +4,7 @@ import { Table, Alert } from 'reactstrap'
 import Pagination from 'components/Common/Pagination'
 import Crowi from 'client/util/Crowi'
 import { Share } from 'client/types/crowi'
-import { formatToLocaleString } from 'client/util/format-date'
+import { formatToLocaleString } from 'client/util/formatDate'
 
 interface Props extends WithTranslation {
   crowi: Crowi
