+++ client/components/Admin/User/UserTableRow.tsx
@@ -3,7 +3,7 @@ import { Badge } from 'reactstrap'
 import UserPicture from 'components/User/UserPicture'
 import UserStatusBadge from 'components/User/UserStatusBadge'
 import UserEditDropdown from './UserEditDropdown'
-import { format } from 'date-fns'
+import format from 'client/util/formatDate'
 
 interface Props {
   me: any
