+++ client/util/formatDate.ts
@@ -1,5 +1,6 @@
 /**
- * OPINIONED date formatter to use browser locale by default
+ * OPINIONED date formatter to use browser locale by default.
+ * Also this treats given string as ISO8601 datetime input. date-fns dropped its behaviour.
  */
 
 import { format as dateFnsFormat, Locale, parseISO, formatDistance as dateFnsFormatDistance } from 'date-fns'
@@ -37,7 +38,6 @@ function sanitizeDateInputForDateFns(input: DateInput) {
 export default function format(date: DateInput, format: string, environment: Environment = defaultEnvironment) {
   return dateFnsFormat(sanitizeDateInputForDateFns(date), format, { locale: environment.getLocale() })
 }
-
 export function formatToLocaleString(date: DateInput, environment: Environment = defaultEnvironment) {
   return format(date, 'PPpp', environment)
 }
@@ -45,7 +45,6 @@ export function formatToLocaleString(date: DateInput, environment: Environment =
 export function formatDistance(date: DateInput | number, base: DateInput, environment: Environment = defaultEnvironment) {
   return dateFnsFormatDistance(sanitizeDateInputForDateFns(date), sanitizeDateInputForDateFns(base), { locale: environment.getLocale(), addSuffix: true })
 }
-
 export function formatDistanceFromNow(date: DateInput, environment: Environment = defaultEnvironment) {
   return formatDistance(date, Date.now(), environment)
 }
